requirements.txt
GitPython==0.3.2.RC1
argparse==1.2.1
async==0.6.1
distribute==0.6.34
feedparser==5.1.3
gitdb==0.5.4
requests==2.3.0
sleekxmpp==1.3.1
smmap==0.8.2
wsgiref==0.1.2
xmltodict